Create HexagonInputElement class
